pyproject.toml
[tool.poetry]
name = "confluence-api"
version = "0.1.0"
description = ""
authors = ["joalvarez <alvarez.joaquinez@gmail.com>"]
readme = "README.md"
packages = [{include = "confluence_api"}]

[tool.poetry.dependencies]
python = "^3.11"
atlassian-python-api = "^3.39.0"
inquirer = "^3.1.3"
python-dotenv = "^1.0.0"
html2markdown = "^0.1.7"
markdownify = "^0.11.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
